chore: move signedtx ext trait
